+++ src/library/lib.rs
@@ -21,7 +21,7 @@ use crate::{
 };
 
 use super::{
-    load::{LibraryLoad, LibraryLoadResult},
+    load::{LibraryLoad, LibraryLoadResult, LoadOpts},
     msg::Message,
     Filter, LibraryUpdate, Song, SongId,
 };
@@ -157,6 +157,7 @@ impl Library {
         &mut self,
         conf: &Config,
         sender: Arc<UnboundedSender<Msg>>,
+        opts: LoadOpts,
     ) -> Result<()> {
         // End panicked processes
         self.any_process();
@@ -168,21 +169,33 @@ impl Library {
         }
 
         let conf = conf.clone();
-        let mut songs = self.songs().clone();
+        let songs = self.songs().clone();
+        let remove_missing =
+            opts.remove_missing.unwrap_or(conf.remove_missing_on_load());
 
         let handle = thread::spawn(move || {
-            let songs = if Self::add_new_songs(&mut songs, &conf) {
-                Some(songs)
-            } else {
-                None
+            let mut res = LibraryLoadResult {
+                removed: false,
+                first_new: songs.len(),
+                add_policy: opts.add_to_playlist,
+                sparse_new: vec![],
+                songs,
             };
+
+            if remove_missing {
+                Self::remove_missing_songs(&mut res);
+            }
+
+            Self::add_new_songs(&mut res, &conf);
 
             if let Err(e) = sender.send(Msg::Library(Message::LoadEnded)) {
                 error!("Library load failed to send message: {e}");
             }
 
-            LibraryLoadResult {
-                new_song_vec: songs,
+            if res.any_change() {
+                Some(res)
+            } else {
+                None
             }
         });
 
@@ -201,6 +214,7 @@ impl Library {
         if let Some(p) = &self.load_process {
             if p.handle.is_finished() {
                 if let Err(e) = self.finish_get_new_songs() {
+                    // TODO: don't ignore Ok()
                     error!("Failed to get new songs: {e}");
                 }
             } else {
@@ -260,8 +274,24 @@ impl UampApp {
     pub fn library_event(&mut self, msg: Message) -> ComMsg<Msg> {
         match msg {
             Message::LoadEnded => {
-                if let Err(e) = self.library.finish_get_new_songs() {
-                    error!("Failed to finsih getting new songs: {e}")
+                match self.library.finish_get_new_songs() {
+                    Err(e) => {
+                        error!("Failed to finsih getting new songs: {e}")
+                    }
+                    Ok(Some(LibraryLoadResult {
+                        first_new,
+                        sparse_new,
+                        add_policy: Some(p),
+                        ..
+                    })) => {
+                        self.player.add_songs(
+                            (first_new..self.library.songs().len())
+                                .map(SongId)
+                                .chain(sparse_new),
+                            p,
+                        );
+                    }
+                    _ => {}
                 }
                 match self
                     .library
@@ -315,14 +345,20 @@ impl Library {
     }
 
     /// Finishes the loading of songs started by `start_get_new_songs`
-    fn finish_get_new_songs(&mut self) -> Result<()> {
+    fn finish_get_new_songs(&mut self) -> Result<Option<LibraryLoadResult>> {
         if let Some(p) = self.load_process.take() {
             let r = p.handle.join().map_err(|_| Error::ThreadPanicked)?;
-            if let Some(s) = r.new_song_vec {
-                *self.songs_mut() = s;
-                self.update(LibraryUpdate::NewData);
+            if let Some(mut s) = r {
+                *self.songs_mut() = mem::take(&mut s.songs);
+                if s.removed {
+                    self.update(LibraryUpdate::RemoveData);
+                } else {
+                    self.update(LibraryUpdate::NewData);
+                }
+                Ok(Some(s))
+            } else {
+                Ok(None)
             }
-            Ok(())
         } else {
             Err(Error::InvalidOperation("No load was running"))
         }
@@ -343,15 +379,47 @@ impl Library {
         }
     }
 
+    fn remove_missing_songs(res: &mut LibraryLoadResult) {
+        let mut remove_any = false;
+        while res
+            .songs
+            .last()
+            .map(|s| {
+                if s.is_deleted() {
+                    true
+                } else if !s.path().exists() {
+                    remove_any = true;
+                    true
+                } else {
+                    false
+                }
+            })
+            .unwrap_or(false)
+        {
+            res.songs.pop();
+        }
+
+        for s in &mut res.songs {
+            if !s.is_deleted() && !s.path().exists() {
+                remove_any = true;
+                s.delete();
+            }
+        }
+
+        res.first_new = res.songs.len();
+        res.removed = remove_any;
+    }
+
     /// Adds new songs to the given vector of songs
-    fn add_new_songs(songs: &mut Vec<Song>, conf: &Config) -> bool {
+    fn add_new_songs(res: &mut LibraryLoadResult, conf: &Config) -> bool {
         let mut new_songs = false;
         let mut paths = conf.search_paths().clone();
         let mut i = 0;
 
-        while songs.last().map(|s| s.is_deleted()).unwrap_or(false) {
-            songs.pop();
+        while res.songs.last().map(|s| s.is_deleted()).unwrap_or(false) {
+            res.songs.pop();
         }
+        res.first_new = res.songs.len();
 
         while i < paths.len() {
             let dir = &paths[i];
@@ -404,7 +472,7 @@ impl Library {
 
                 let mut idx = None;
 
-                for (i, s) in songs.iter().enumerate() {
+                for (i, s) in res.songs.iter().enumerate() {
                     if s.is_deleted() {
                         // prefer the later indexes, user is more likely to
                         // remove old song and songs at the end are more esily
@@ -420,9 +488,10 @@ impl Library {
 
                 if let Ok(song) = Song::from_path(path) {
                     if let Some(i) = idx {
-                        songs[i] = song;
+                        res.sparse_new.push(SongId(i));
+                        res.songs[i] = song;
                     } else {
-                        songs.push(song);
+                        res.songs.push(song);
                     }
                 }
             }