+++ src/player/add_policy.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Copy, Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub enum AddPolicy {
+    End,
+    Next,
+    MixIn,
+}